+++ resources/lang/en/privacy-blur.php
@@ -1,0 +1,6 @@
+<?php
+
+// translations for Arseno25/FilamentPrivacyBlur
+return [
+    //
+];