+++ src/Facades/FilamentPrivacyBlur.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Arseno25\FilamentPrivacyBlur\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \Arseno25\FilamentPrivacyBlur\FilamentPrivacyBlur
+ */
+class FilamentPrivacyBlur extends Facade
+{
+    protected static function getFacadeAccessor(): string
+    {
+        return \Arseno25\FilamentPrivacyBlur\FilamentPrivacyBlur::class;
+    }
+}