+++ src/FilamentPrivacyBlur.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Arseno25\FilamentPrivacyBlur;
+
+class FilamentPrivacyBlur {}