+++ src/Models/PrivacyRevealLog.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Arseno25\FilamentPrivacyBlur\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class PrivacyRevealLog extends Model
+{
+    protected $table = 'privacy_reveal_logs';
+
+    protected $guarded = [];
+}